fix: reuse canonical review marker parser
