+++ internal/reviewer/runner_test.go
@@ -1559,388 +1559,6 @@ func TestDiscoverPullRequestsRequeuesFollowUpOnNewHeadWithoutFreshReviewRequest(
 	}
 }
 
-func TestDiscoverPullRequestsRequeuesLooperEngagedFollowUpWithoutPublishedHeadMetadata(t *testing.T) {
-	t.Parallel()
-	fixture := newRunnerFixture(t)
-	const loopID = "loop_follow_engaged_without_metadata"
-	github := &fakeGitHubGateway{
-		currentLogin:   "bob",
-		reviewRequests: []string{},
-		viewHeadSHA:    "new-head",
-		reviews: []map[string]any{{
-			"author": map[string]any{"login": "bob"},
-			"commit": map[string]any{"oid": "old-head"},
-			"state":  "CHANGES_REQUESTED",
-			"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=blocking -->",
-		}},
-	}
-	agent := &fakeAgentExecutor{results: []AgentResult{{Status: "completed", Summary: "No actionable findings", Stdout: `__LOOPER_RESULT__={"summary":"posted clean follow-up review"}`, ParseStatus: "parsed"}}}
-	runner := New(Options{DB: fixture.coordinator.DB(), Repos: fixture.repos, GitHub: github, Git: &fakeGitGateway{}, AgentExecutor: agent, Logger: fixture.logger, Now: fixture.now, DiscoveryPolicy: DiscoveryPolicy{AutoDiscovery: true, IncludeDrafts: false, RequireReviewRequest: true, Labels: []string{}, LabelMode: config.LabelModeAll}, ReviewEvents: config.ReviewerReviewEventsConfig{Clean: config.ReviewerReviewEventComment, Blocking: config.ReviewerReviewEventRequestChanges}, LoopConfig: testReviewerLoopConfig()})
-	nowISO := fixture.nowISO()
-	repo := "acme/looper"
-	prNumber := int64(42)
-	metadata := `{"followUpdates":true,"lastFilterSkip":{"kind":"not_requested","headSha":"old-head","reviewerLogin":"bob"},"loop":{"enabled":true,"iterationCount":1,"iterationsByHead":{"old-head":1}}}`
-	loop := storage.LoopRecord{ID: loopID, Seq: 1, ProjectID: "project_1", Type: "reviewer", TargetType: "pull_request", Repo: &repo, PRNumber: &prNumber, Status: "completed", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := fixture.repos.Loops.Upsert(context.Background(), loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	result, err := runner.DiscoverPullRequests(context.Background(), DiscoveryInput{ProjectID: "project_1", Repo: repo})
-	if err != nil {
-		t.Fatalf("DiscoverPullRequests() error = %v", err)
-	}
-	if len(result.QueueItems) != 1 {
-		t.Fatalf("len(QueueItems) = %d, want 1 engaged follow-up", len(result.QueueItems))
-	}
-	persistedLoop, err := fixture.repos.Loops.GetByID(context.Background(), loop.ID)
-	if err != nil {
-		t.Fatalf("Loops.GetByID() error = %v", err)
-	}
-	if persistedLoop == nil || persistedLoop.MetadataJSON == nil || !contains(*persistedLoop.MetadataJSON, `"lastPublishedHeadSha":"old-head"`) {
-		t.Fatalf("loop after discovery = %#v, want GitHub engagement head backfilled", persistedLoop)
-	}
-	fixture.advance(time.Hour)
-	claimed, err := fixture.repos.Queue.ClaimNextOfType(context.Background(), fixture.nowISO(), "reviewer-worker-1", "reviewer")
-	if err != nil || claimed == nil {
-		t.Fatalf("ClaimNextOfType() = (%#v, %v), want engaged follow-up", claimed, err)
-	}
-	processed, err := runner.ProcessClaimedItem(context.Background(), *claimed)
-	if err != nil {
-		t.Fatalf("ProcessClaimedItem() error = %v", err)
-	}
-	if processed.Status != "success" || len(agent.starts) != 1 {
-		t.Fatalf("ProcessClaimedItem() = %#v, agent starts = %d; want full follow-up review", processed, len(agent.starts))
-	}
-}
-
-func TestDiscoverPullRequestsBackfillsEngagementWhenDiscoveryViewOmitsReviews(t *testing.T) {
-	t.Parallel()
-	fixture := newRunnerFixture(t)
-	const loopID = "loop_follow_engaged_snapshot_omits_reviews"
-	// Simulate daemon DiscoverySnapshot lifecycle: discovery ViewPullRequest uses a
-	// fixer-profile detail without reviews, so engagement recovery must load
-	// reviews via LoadPullRequestReviews rather than trusting detail.Reviews.
-	github := &fakeGitHubGateway{
-		currentLogin:      "bob",
-		reviewRequests:    []string{},
-		viewHeadSHA:       "new-head",
-		omitReviewsOnView: true,
-		reviews: []map[string]any{{
-			"author": map[string]any{"login": "bob"},
-			"commit": map[string]any{"oid": "old-head"},
-			"state":  "CHANGES_REQUESTED",
-			"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=blocking -->",
-		}},
-	}
-	agent := &fakeAgentExecutor{results: []AgentResult{{Status: "completed", Summary: "No actionable findings", Stdout: `__LOOPER_RESULT__={"summary":"posted clean follow-up review"}`, ParseStatus: "parsed"}}}
-	runner := New(Options{DB: fixture.coordinator.DB(), Repos: fixture.repos, GitHub: github, Git: &fakeGitGateway{}, AgentExecutor: agent, Logger: fixture.logger, Now: fixture.now, DiscoveryPolicy: DiscoveryPolicy{AutoDiscovery: true, IncludeDrafts: false, RequireReviewRequest: true, Labels: []string{}, LabelMode: config.LabelModeAll}, ReviewEvents: config.ReviewerReviewEventsConfig{Clean: config.ReviewerReviewEventComment, Blocking: config.ReviewerReviewEventRequestChanges}, LoopConfig: testReviewerLoopConfig()})
-	nowISO := fixture.nowISO()
-	repo := "acme/looper"
-	prNumber := int64(42)
-	metadata := `{"followUpdates":true,"lastFilterSkip":{"kind":"not_requested","headSha":"old-head","reviewerLogin":"bob"},"loop":{"enabled":true,"iterationCount":1,"iterationsByHead":{"old-head":1}}}`
-	loop := storage.LoopRecord{ID: loopID, Seq: 1, ProjectID: "project_1", Type: "reviewer", TargetType: "pull_request", Repo: &repo, PRNumber: &prNumber, Status: "completed", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := fixture.repos.Loops.Upsert(context.Background(), loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	// Attach a non-nil snapshot the same way the scheduler tick does, so this
-	// covers the daemon discovery lifecycle rather than only direct runner use.
-	// The snapshot gateway is unused by the fake adapter; omitReviewsOnView is
-	// what models the fixer-profile empty-reviews detail shape.
-	snapshotGateway := githubinfra.New(githubinfra.Options{GHRun: func(context.Context, shell.Options) (shell.Result, error) {
-		return shell.Result{}, errors.New("discovery snapshot gateway should not be called by fake reviewer path")
-	}})
-	snapshot := githubinfra.NewDiscoverySnapshot(snapshotGateway, githubinfra.NewDiscoveryTickState(), githubinfra.DiscoverySnapshotOptions{})
-	if snapshot == nil {
-		t.Fatal("NewDiscoverySnapshot() = nil, want non-nil scheduler-style snapshot")
-	}
-	result, err := runner.DiscoverPullRequests(context.Background(), DiscoveryInput{ProjectID: "project_1", Repo: repo, Snapshot: snapshot})
-	if err != nil {
-		t.Fatalf("DiscoverPullRequests() error = %v", err)
-	}
-	if len(result.QueueItems) != 1 {
-		t.Fatalf("len(QueueItems) = %d, want 1 engaged follow-up after review reload", len(result.QueueItems))
-	}
-	if github.loadReviewsCalls < 1 {
-		t.Fatalf("LoadPullRequestReviews calls = %d, want >= 1 when discovery view omits reviews", github.loadReviewsCalls)
-	}
-	persistedLoop, err := fixture.repos.Loops.GetByID(context.Background(), loop.ID)
-	if err != nil {
-		t.Fatalf("Loops.GetByID() error = %v", err)
-	}
-	if persistedLoop == nil || persistedLoop.MetadataJSON == nil || !contains(*persistedLoop.MetadataJSON, `"lastPublishedHeadSha":"old-head"`) {
-		t.Fatalf("loop after discovery = %#v, want engagement head backfilled from reloaded reviews", persistedLoop)
-	}
-	fixture.advance(time.Hour)
-	claimed, err := fixture.repos.Queue.ClaimNextOfType(context.Background(), fixture.nowISO(), "reviewer-worker-1", "reviewer")
-	if err != nil || claimed == nil {
-		t.Fatalf("ClaimNextOfType() = (%#v, %v), want engaged follow-up", claimed, err)
-	}
-	processed, err := runner.ProcessClaimedItem(context.Background(), *claimed)
-	if err != nil {
-		t.Fatalf("ProcessClaimedItem() error = %v", err)
-	}
-	if processed.Status != "success" || len(agent.starts) != 1 {
-		t.Fatalf("ProcessClaimedItem() = %#v, agent starts = %d; want full follow-up review", processed, len(agent.starts))
-	}
-}
-
-func TestDiscoverPullRequestsRejectsPolicyInconsistentEngagementMarkers(t *testing.T) {
-	t.Parallel()
-	// Publish verification rejects COMMENTED reviews that carry clean/blocking
-	// when policy requires APPROVE/REQUEST_CHANGES. Engagement backfill must
-	// apply the same outcome/event/policy gate so those rejected markers cannot
-	// restore lastPublishedHeadSha and bypass requireReviewRequest.
-	cases := []struct {
-		name    string
-		state   string
-		outcome string
-		policy  config.ReviewerReviewEventsConfig
-	}{
-		{
-			name:    "commented_blocking_under_request_changes_policy",
-			state:   "COMMENTED",
-			outcome: "blocking",
-			policy:  config.ReviewerReviewEventsConfig{Clean: config.ReviewerReviewEventComment, Blocking: config.ReviewerReviewEventRequestChanges},
-		},
-		{
-			name:    "commented_clean_under_approve_policy",
-			state:   "COMMENTED",
-			outcome: "clean",
-			policy:  config.ReviewerReviewEventsConfig{Clean: config.ReviewerReviewEventApprove, Blocking: config.ReviewerReviewEventComment},
-		},
-	}
-	for _, tc := range cases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			fixture := newRunnerFixture(t)
-			loopID := "loop_reject_inconsistent_" + tc.name
-			github := &fakeGitHubGateway{
-				currentLogin:   "bob",
-				reviewRequests: []string{},
-				viewHeadSHA:    "new-head",
-				reviews: []map[string]any{{
-					"author": map[string]any{"login": "bob"},
-					"commit": map[string]any{"oid": "old-head"},
-					"state":  tc.state,
-					"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=" + tc.outcome + " -->",
-				}},
-			}
-			runner := New(Options{
-				DB: fixture.coordinator.DB(), Repos: fixture.repos, GitHub: github, Git: &fakeGitGateway{},
-				AgentExecutor: &fakeAgentExecutor{}, Logger: fixture.logger, Now: fixture.now,
-				DiscoveryPolicy: DiscoveryPolicy{AutoDiscovery: true, IncludeDrafts: false, RequireReviewRequest: true, Labels: []string{}, LabelMode: config.LabelModeAll},
-				ReviewEvents:    tc.policy,
-				LoopConfig:      testReviewerLoopConfig(),
-			})
-			nowISO := fixture.nowISO()
-			repo := "acme/looper"
-			prNumber := int64(42)
-			metadata := `{"followUpdates":true,"lastFilterSkip":{"kind":"not_requested","headSha":"old-head","reviewerLogin":"bob"},"loop":{"enabled":true,"iterationCount":1,"iterationsByHead":{"old-head":1}}}`
-			loop := storage.LoopRecord{ID: loopID, Seq: 1, ProjectID: "project_1", Type: "reviewer", TargetType: "pull_request", Repo: &repo, PRNumber: &prNumber, Status: "completed", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO}
-			if err := fixture.repos.Loops.Upsert(context.Background(), loop); err != nil {
-				t.Fatalf("Loops.Upsert() error = %v", err)
-			}
-
-			result, err := runner.DiscoverPullRequests(context.Background(), DiscoveryInput{ProjectID: "project_1", Repo: repo})
-			if err != nil {
-				t.Fatalf("DiscoverPullRequests() error = %v", err)
-			}
-			if len(result.QueueItems) != 0 {
-				t.Fatalf("len(QueueItems) = %d, want 0 (policy-inconsistent marker must not authorize follow-up)", len(result.QueueItems))
-			}
-			persistedLoop, err := fixture.repos.Loops.GetByID(context.Background(), loop.ID)
-			if err != nil {
-				t.Fatalf("Loops.GetByID() error = %v", err)
-			}
-			if persistedLoop == nil || persistedLoop.MetadataJSON == nil {
-				t.Fatalf("loop after discovery = %#v, want preserved metadata", persistedLoop)
-			}
-			if contains(*persistedLoop.MetadataJSON, `"lastPublishedHeadSha"`) {
-				t.Fatalf("loop metadata = %s, want no lastPublishedHeadSha backfill from rejected marker", *persistedLoop.MetadataJSON)
-			}
-		})
-	}
-}
-
-func TestLooperReviewEngagementHeadRequiresOutcomeEventPolicyConsistency(t *testing.T) {
-	t.Parallel()
-	const loopID = "loop_engagement_policy"
-	reviews := []map[string]any{{
-		"author": map[string]any{"login": "bob"},
-		"commit": map[string]any{"oid": "old-head"},
-		"state":  "COMMENTED",
-		"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=blocking -->",
-	}}
-	allowedWithRequestChanges := []ReviewEvent{ReviewEventComment, ReviewEventRequestChanges}
-	if got := looperReviewEngagementHead(reviews, "bob", loopID, "new-head", allowedWithRequestChanges, false); got != "" {
-		t.Fatalf("COMMENTED+blocking under REQUEST_CHANGES policy = %q, want empty", got)
-	}
-	allowedCommentOnly := []ReviewEvent{ReviewEventComment}
-	if got := looperReviewEngagementHead(reviews, "bob", loopID, "new-head", allowedCommentOnly, false); got != "old-head" {
-		t.Fatalf("COMMENTED+blocking under COMMENT policy = %q, want old-head", got)
-	}
-	requestChangesReviews := []map[string]any{{
-		"author": map[string]any{"login": "bob"},
-		"commit": map[string]any{"oid": "old-head"},
-		"state":  "CHANGES_REQUESTED",
-		"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=blocking -->",
-	}}
-	if got := looperReviewEngagementHead(requestChangesReviews, "bob", loopID, "new-head", allowedWithRequestChanges, false); got != "old-head" {
-		t.Fatalf("CHANGES_REQUESTED+blocking under REQUEST_CHANGES policy = %q, want old-head", got)
-	}
-	cleanCommented := []map[string]any{{
-		"author": map[string]any{"login": "bob"},
-		"commit": map[string]any{"oid": "old-head"},
-		"state":  "COMMENTED",
-		"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=clean -->",
-	}}
-	allowedWithApprove := []ReviewEvent{ReviewEventComment, ReviewEventApprove}
-	if got := looperReviewEngagementHead(cleanCommented, "bob", loopID, "new-head", allowedWithApprove, false); got != "" {
-		t.Fatalf("COMMENTED+clean under APPROVE policy without self-approval = %q, want empty", got)
-	}
-	if got := looperReviewEngagementHead(cleanCommented, "bob", loopID, "new-head", allowedWithApprove, true); got != "old-head" {
-		t.Fatalf("COMMENTED+clean under APPROVE policy with self-approval fallback = %q, want old-head", got)
-	}
-}
-
-func TestLooperReviewEngagementHeadRejectsNonCanonicalOutcomeCasing(t *testing.T) {
-	t.Parallel()
-	// Publish verification (isValidReviewMarkerOutcome / runtimeValidReviewMarkerOutcome)
-	// accepts only exact lowercase outcomes. Case variants must not authorize
-	// lastPublishedHeadSha backfill and bypass requireReviewRequest.
-	const loopID = "loop_engagement_outcome_case"
-	allowed := []ReviewEvent{ReviewEventComment, ReviewEventRequestChanges}
-	for _, outcome := range []string{"BLOCKING", "Blocking", "CLEAN", "NON_BLOCKING", "Actionable"} {
-		reviews := []map[string]any{{
-			"author": map[string]any{"login": "bob"},
-			"commit": map[string]any{"oid": "old-head"},
-			"state":  "CHANGES_REQUESTED",
-			"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=" + outcome + " -->",
-		}}
-		if got := looperReviewEngagementHead(reviews, "bob", loopID, "new-head", allowed, false); got != "" {
-			t.Fatalf("outcome=%q engagement head = %q, want empty (publish verification would reject)", outcome, got)
-		}
-		if reviewMarkerOutcomeEventAllowed(outcome, ReviewEventRequestChanges, allowed, false) {
-			t.Fatalf("reviewMarkerOutcomeEventAllowed(%q) = true, want false", outcome)
-		}
-	}
-	if !isRecognizedReviewOutcome("blocking") || isRecognizedReviewOutcome("BLOCKING") {
-		t.Fatal("isRecognizedReviewOutcome must match exact lowercase publish-verification tokens")
-	}
-}
-
-func TestDiscoverPullRequestsRejectsNonCanonicalOutcomeCasingEngagement(t *testing.T) {
-	t.Parallel()
-	fixture := newRunnerFixture(t)
-	const loopID = "loop_reject_outcome_BLOCKING"
-	github := &fakeGitHubGateway{
-		currentLogin:   "bob",
-		reviewRequests: []string{},
-		viewHeadSHA:    "new-head",
-		reviews: []map[string]any{{
-			"author": map[string]any{"login": "bob"},
-			"commit": map[string]any{"oid": "old-head"},
-			"state":  "CHANGES_REQUESTED",
-			"body":   "<!-- looper:review id=reviewer:" + loopID + " head=old-head outcome=BLOCKING -->",
-		}},
-	}
-	runner := New(Options{
-		DB: fixture.coordinator.DB(), Repos: fixture.repos, GitHub: github, Git: &fakeGitGateway{},
-		AgentExecutor: &fakeAgentExecutor{}, Logger: fixture.logger, Now: fixture.now,
-		DiscoveryPolicy: DiscoveryPolicy{AutoDiscovery: true, IncludeDrafts: false, RequireReviewRequest: true, Labels: []string{}, LabelMode: config.LabelModeAll},
-		ReviewEvents:    config.ReviewerReviewEventsConfig{Clean: config.ReviewerReviewEventComment, Blocking: config.ReviewerReviewEventRequestChanges},
-		LoopConfig:      testReviewerLoopConfig(),
-	})
-	nowISO := fixture.nowISO()
-	repo := "acme/looper"
-	prNumber := int64(42)
-	metadata := `{"followUpdates":true,"lastFilterSkip":{"kind":"not_requested","headSha":"old-head","reviewerLogin":"bob"},"loop":{"enabled":true,"iterationCount":1,"iterationsByHead":{"old-head":1}}}`
-	loop := storage.LoopRecord{ID: loopID, Seq: 1, ProjectID: "project_1", Type: "reviewer", TargetType: "pull_request", Repo: &repo, PRNumber: &prNumber, Status: "completed", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := fixture.repos.Loops.Upsert(context.Background(), loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	result, err := runner.DiscoverPullRequests(context.Background(), DiscoveryInput{ProjectID: "project_1", Repo: repo})
-	if err != nil {
-		t.Fatalf("DiscoverPullRequests() error = %v", err)
-	}
-	if len(result.QueueItems) != 0 {
-		t.Fatalf("len(QueueItems) = %d, want 0 (outcome=BLOCKING must not authorize follow-up)", len(result.QueueItems))
-	}
-	persistedLoop, err := fixture.repos.Loops.GetByID(context.Background(), loop.ID)
-	if err != nil {
-		t.Fatalf("Loops.GetByID() error = %v", err)
-	}
-	if persistedLoop == nil || persistedLoop.MetadataJSON == nil {
-		t.Fatalf("loop after discovery = %#v, want preserved metadata", persistedLoop)
-	}
-	if contains(*persistedLoop.MetadataJSON, `"lastPublishedHeadSha"`) {
-		t.Fatalf("loop metadata = %s, want no lastPublishedHeadSha backfill from case-variant outcome", *persistedLoop.MetadataJSON)
-	}
-}
-
-func TestReviewsForEngagementBackfillTreatsEmptyPresentListAsAuthoritative(t *testing.T) {
-	t.Parallel()
-	fixture := newRunnerFixture(t)
-	const loopID = "loop_empty_reviews_no_reload"
-	github := &fakeGitHubGateway{
-		currentLogin:   "bob",
-		reviewRequests: []string{},
-		viewHeadSHA:    "new-head",
-		// Present but empty reviews list (reviewer profile with no submissions).
-		reviews: []map[string]any{},
-	}
-	runner := New(Options{
-		DB: fixture.coordinator.DB(), Repos: fixture.repos, GitHub: github, Git: &fakeGitGateway{},
-		AgentExecutor: &fakeAgentExecutor{}, Logger: fixture.logger, Now: fixture.now,
-		DiscoveryPolicy: DiscoveryPolicy{AutoDiscovery: true, IncludeDrafts: false, RequireReviewRequest: true, Labels: []string{}, LabelMode: config.LabelModeAll},
-		ReviewEvents:    config.ReviewerReviewEventsConfig{Clean: config.ReviewerReviewEventComment, Blocking: config.ReviewerReviewEventRequestChanges},
-		LoopConfig:      testReviewerLoopConfig(),
-	})
-	nowISO := fixture.nowISO()
-	repo := "acme/looper"
-	prNumber := int64(42)
-	metadata := `{"followUpdates":true,"loop":{"enabled":true,"iterationCount":1,"iterationsByHead":{"old-head":1}}}`
-	loop := storage.LoopRecord{ID: loopID, Seq: 1, ProjectID: "project_1", Type: "reviewer", TargetType: "pull_request", Repo: &repo, PRNumber: &prNumber, Status: "completed", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := fixture.repos.Loops.Upsert(context.Background(), loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	project, err := fixture.repos.Projects.GetByID(context.Background(), "project_1")
-	if err != nil || project == nil {
-		t.Fatalf("Projects.GetByID() = (%#v, %v)", project, err)
-	}
-
-	// Empty but non-nil reviews must not trigger LoadPullRequestReviews.
-	got, err := runner.reviewsForEngagementBackfill(context.Background(), *project, loop, PullRequestDetail{
-		Number: prNumber, HeadSHA: "new-head", Reviews: []map[string]any{},
-	})
-	if err != nil {
-		t.Fatalf("reviewsForEngagementBackfill(empty present) error = %v", err)
-	}
-	if got == nil || len(got) != 0 {
-		t.Fatalf("reviewsForEngagementBackfill(empty present) = %#v, want empty non-nil slice", got)
-	}
-	if github.loadReviewsCalls != 0 {
-		t.Fatalf("LoadPullRequestReviews calls = %d, want 0 when reviews list is present and empty", github.loadReviewsCalls)
-	}
-
-	// Omitted (nil) reviews still load via the recovery path.
-	got, err = runner.reviewsForEngagementBackfill(context.Background(), *project, loop, PullRequestDetail{
-		Number: prNumber, HeadSHA: "new-head", Reviews: nil,
-	})
-	if err != nil {
-		t.Fatalf("reviewsForEngagementBackfill(nil) error = %v", err)
-	}
-	if got == nil {
-		t.Fatal("reviewsForEngagementBackfill(nil) = nil, want loaded empty list from gateway")
-	}
-	if github.loadReviewsCalls != 1 {
-		t.Fatalf("LoadPullRequestReviews calls = %d, want 1 when reviews field is omitted", github.loadReviewsCalls)
-	}
-}
-
 func TestDiscoverPullRequestsSkipsDraftFollowUpWithoutTerminatingLoop(t *testing.T) {
 	t.Parallel()
 	fixture := newRunnerFixture(t)
